requirements.txt
pydantic-ai
python-dotenv
groq
pytest
pytest-asyncio
requests